general fix

+++ controllers/basics-controller.js
@@ -6,6 +6,7 @@ const User = require('../models/user');
 
 const getBasics = async (req, res, next) => {
     const patientId = req.params.pid;
+    console.log('patientId:', patientId)
     let patient;
     try {
         patient = await Patient.findById(patientId);
@@ -13,21 +14,21 @@ const getBasics = async (req, res, next) => {
         return next(new HttpError('Κάτι πήγε λάθος,δεν φορτώθηκαν τα στοιχεία του ασθενή.', 500));
     }
     res.json(patient)
-   
+
 };
 const createImageBasics = async (req, res, next) => {
     const myId = mongoose.Types.ObjectId();
     const { patientId, sirname, name, amka, tel, dateOfBirth, placeOfBirth, job, familyStatus, gender, address, area, postalCode, email, fathersName, uid } = req.body;
-    const file=req.file.path;
+    const file = req.file.path;
     let patient;
-    
+
     try {
         patient = await Patient.findOne({ amka: amka });
     } catch (err) {
         return next(new HttpError('Κάτι πήγε λάθος,παρακαλώ προσπαθήστε ξανά.'));
     }
     if (patient && (patientId === null)) {
-        return next(new HttpError('Υπάρχει ήδη καταχωρημένος αυτός ο ασθενής.'));
+        return next(new HttpError('Υπάρχει ήδη καταχωρημένος ασθενής με αυτό το ΑΜΚΑ.'));
     } else {
         const createdBasics = new Patient({
             _id: myId,
@@ -86,14 +87,14 @@ const createBasics = async (req, res, next) => {
     const myId = mongoose.Types.ObjectId();
     const { patientId, sirname, name, amka, tel, dateOfBirth, placeOfBirth, job, familyStatus, gender, address, area, postalCode, email, fathersName, uid } = req.body;
     let patient;
-   
+
     try {
         patient = await Patient.findOne({ amka: amka });
     } catch (err) {
         return next(new HttpError('Κάτι πήγε λάθος,παρακαλώ προσπαθήστε ξανά.'));
     }
     if (patient && (patientId === null)) {
-        return next(new HttpError('Υπάρχει ήδη καταχωρημένος αυτός ο ασθενής.'));
+        return next(new HttpError('Υπάρχει ήδη καταχωρημένος ασθενής με αυτό το ΑΜΚΑ.'));
     } else {
         const createdBasics = new Patient({
             _id: myId,
@@ -114,7 +115,7 @@ const createBasics = async (req, res, next) => {
             doctor: uid,
             basic: null,
             anamnistiko: null,
-           
+
             visits: []
         });
         let doctor;
@@ -150,20 +151,23 @@ const createBasics = async (req, res, next) => {
 
 
 const updateImageBasics = async (req, res, next) => {
-    const {  sirname, name, amka, tel, dateOfBirth, placeOfBirth, job, familyStatus, gender, address, area, postalCode, email, fathersName } = req.body;
+    console.log('here')
+    const { sirname, name, amka, tel, dateOfBirth, placeOfBirth, job, familyStatus, gender, address, area, postalCode, email, fathersName } = req.body;
     let patient;
-    const patientId=req.params.pid
-    const file=req.file.path;
-   
+    const patientId = req.params.pid
+    const file = req.file.path;
+
     try {
         patient = await Patient.findById(patientId);
     } catch (err) {
         return next(new HttpError('Κάτι πήγε λάθος,δεν βρέθηκε αυτός ο ασθενής.'));
     }
-   
-    fs.unlink(`uploads/images/${patient.files.split('\\')[2]}`, (err) => {
-        console.log(err)
-    });
+
+    if (!!patient.files) {
+        fs.unlink(`uploads/images/${patient.files.split('\\')[2]}`, (err) => {
+            console.log(err)
+        });
+    }
 
 
     patient.sirname = sirname
@@ -180,7 +184,7 @@ const updateImageBasics = async (req, res, next) => {
     patient.area = area
     patient.postalCode = postalCode
     patient.email = email
-    patient.files=file;
+    patient.files = file;
 
 
     try {
@@ -195,10 +199,10 @@ const updateImageBasics = async (req, res, next) => {
 }
 
 const updateBasics = async (req, res, next) => {
-    const {  sirname, name, amka, tel, dateOfBirth, placeOfBirth, job, familyStatus, gender, address, area, postalCode, email, fathersName } = req.body;
+    const { sirname, name, amka, tel, dateOfBirth, placeOfBirth, job, familyStatus, gender, address, area, postalCode, email, fathersName } = req.body;
     let patient;
-    const patientId=req.params.pid
-   
+    const patientId = req.params.pid
+
     try {
         patient = await Patient.findById(patientId);
     } catch (err) {
@@ -232,8 +236,8 @@ const updateBasics = async (req, res, next) => {
 
 }
 
-exports.createImageBasics=createImageBasics;
+exports.createImageBasics = createImageBasics;
 exports.getBasics = getBasics;
 exports.createBasics = createBasics;
 exports.updateBasics = updateBasics;
-exports.updateImageBasics=updateImageBasics;
+exports.updateImageBasics = updateImageBasics;
